+++ client/src/pages/EventPage/AdminPage/EventLoginPage.tsx
@@ -1,38 +1,12 @@
-import {useState} from 'react';
-import {FixedButton, MainLayout, LabelInput, Title} from 'haengdong-design';
+import {FixedButton, LabelInput, Title} from 'haengdong-design';
 
-import validateEventPassword from '@utils/validate/validateEventPassword';
-import useRequestPostLogin from '@hooks/queries/useRequestPostLogin';
+import useEventLogin from '@hooks/useEventLogin';
 
 import RULE from '@constants/rule';
 import {PASSWORD_LENGTH} from '@constants/password';
 
 const EventLoginPage = () => {
-  const [password, setPassword] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
-  const [canSubmit, setCanSubmit] = useState(false);
-  const {mutate: postLogin} = useRequestPostLogin();
-
-  const submitPassword = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    postLogin({password}, {onError: () => setErrorMessage('비밀번호가 틀렸어요')});
-  };
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = event.target.value;
-    const validation = validateEventPassword(newValue);
-
-    setCanSubmit(newValue.length === RULE.maxEventPasswordLength);
-
-    if (validation.isValid) {
-      setPassword(newValue);
-      setErrorMessage('');
-    } else {
-      event.target.value = password;
-      setErrorMessage(validation.errorMessage ?? '');
-    }
-  };
+  const {password, errorMessage, handleChange, canSubmit, submitPassword} = useEventLogin();
 
   return (
     <>
